+++ mockserver-core/src/main/java/org/mockserver/proxyconfiguration/ProxyConfiguration.java
@@ -8,9 +8,7 @@ import org.mockserver.model.ObjectWithJsonToString;
 import java.net.InetSocketAddress;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import static io.netty.handler.codec.http.HttpHeaderNames.PROXY_AUTHORIZATION;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
@@ -40,25 +38,16 @@ public class ProxyConfiguration extends ObjectWithJsonToString {
         String password = forwardProxyAuthenticationPassword();
 
         InetSocketAddress httpProxySocketAddress = forwardHttpProxy();
-        if (httpProxySocketAddress == null) {
-            httpProxySocketAddress = httpProxy();
-        }
         if (httpProxySocketAddress != null) {
             proxyConfigurations.add(proxyConfiguration(Type.HTTP, httpProxySocketAddress, username, password));
         }
 
         InetSocketAddress httpsProxySocketAddress = forwardHttpsProxy();
-        if (httpsProxySocketAddress == null) {
-            httpsProxySocketAddress = httpsProxy();
-        }
         if (httpsProxySocketAddress != null) {
             proxyConfigurations.add(proxyConfiguration(Type.HTTPS, httpsProxySocketAddress, username, password));
         }
 
         InetSocketAddress socksProxySocketAddress = forwardSocksProxy();
-        if (socksProxySocketAddress == null) {
-            socksProxySocketAddress = socksProxy();
-        }
         if (socksProxySocketAddress != null) {
             if (proxyConfigurations.isEmpty()) {
                 proxyConfigurations.add(proxyConfiguration(Type.SOCKS5, socksProxySocketAddress, username, password));
@@ -115,8 +104,8 @@ public class ProxyConfiguration extends ObjectWithJsonToString {
     public ProxyConfiguration addProxyAuthenticationHeader(HttpRequest httpRequest) {
         if (isNotBlank(username) && isNotBlank(password)) {
             httpRequest.withHeader(
-                    PROXY_AUTHORIZATION.toString(),
-                    "Basic " + Base64.encode(Unpooled.copiedBuffer(username + ':' + password, StandardCharsets.UTF_8), false).toString(StandardCharsets.US_ASCII)
+                PROXY_AUTHORIZATION.toString(),
+                "Basic " + Base64.encode(Unpooled.copiedBuffer(username + ':' + password, StandardCharsets.UTF_8), false).toString(StandardCharsets.US_ASCII)
             );
         }
         return this;